requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & AI
openai==1.10.0

# Vector Database
chromadb==0.4.22

# Data Processing
python-dateutil==2.8.2
pytz==2024.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Data Validation
email-validator==2.1.0

# JSON Processing
orjson==3.9.12